new cover API in the EkoOptions
