+++ samples/helloeko/src/main/java/com/eko/sdktest/CustomCover.kt
@@ -1,0 +1,11 @@
+package com.eko.sdktest
+
+import android.content.Context
+import android.graphics.Color
+import android.view.View
+
+class CustomCover(context: Context?) : View(context) {
+    init {
+        setBackgroundColor(Color.BLUE)
+    }
+}
